break aggregated_rp into multiple files
