+++ src/aggregated_range_proof/mod.rs
@@ -1,0 +1,7 @@
+#![allow(non_snake_case)]
+
+mod party;
+mod dealer;
+mod messages;
+
+
